Rename syntactic_checker package to syntactic
