+++ tests/output/src/main/scala/example/Output.scala
@@ -1,5 +1,5 @@
 package output.example
 
-import syntactic_checker.blacklist.BlacklistChecker
+import syntactic.blacklist.BlacklistChecker
 
 case class Output(checker: BlacklistChecker, expected: List[(Int, Int)])
